+++ src/main/java/entities/Friend.java
@@ -4,8 +4,8 @@ import java.sql.Timestamp;
 
 public class Friend {
     private Timestamp tmp;
-    private Integer user_1;
-    private Integer user_2;
+    private Long user_1;
+    private Long user_2;
 
     public Timestamp getTmp() {
         return tmp;
@@ -15,19 +15,21 @@ public class Friend {
         this.tmp = tmp;
     }
 
-    public Integer getUser_1() {
+
+    public Long getUser_1() {
         return user_1;
     }
 
-    public void setUser_1(Integer user_1) {
+    public void setUser_1(Long user_1) {
         this.user_1 = user_1;
     }
 
-    public Integer getUser_2() {
+
+    public Long getUser_2() {
         return user_2;
     }
 
-    public void setUser_2(Integer user_2) {
+    public void setUser_2(Long user_2) {
         this.user_2 = user_2;
     }
 }
