+++ src/main/java/entities/Message.java
@@ -1,21 +1,20 @@
 package entities;
 
 import java.io.Serializable;
-import java.sql.Timestamp;
 
 public class Message implements Serializable {
 
     private Integer type;
     private String tmp;
-    private Integer user_id1;
+    private Long user_id1;
     private String user_name;
     private Long comment_id;
     private String comment;
-    private Integer post_id;
+    private Long post_id;
     private String post;
-    private Integer user_id2;
+    private Long user_id2;
     private Long comment_replied;
-    private Integer post_commented;
+    private Long post_commented;
 
     public Long getComment_replied() {
         return comment_replied;
@@ -25,17 +24,18 @@ public class Message implements Serializable {
         this.comment_replied = comment_replied;
     }
 
-    public Integer getPost_commented() {
+    public Long getPost_commented() {
         return post_commented;
     }
 
-    public void setPost_commented(Integer post_commented) {
+    public void setPost_commented(Long post_commented) {
         this.post_commented = post_commented;
     }
 
     public Message (Integer type){
         this.type = type;
     }
+
 
     public Integer getType() {
         return type;
@@ -45,6 +45,7 @@ public class Message implements Serializable {
         this.type = type;
     }
 
+
     public String getTmp() {
         return tmp;
     }
@@ -53,13 +54,15 @@ public class Message implements Serializable {
         this.tmp = tmp;
     }
 
-    public Integer getUser_id1() {
+
+    public Long getUser_id1() {
         return user_id1;
     }
 
-    public void setUser_id1(Integer user_id) {
+    public void setUser_id1(Long user_id) {
         this.user_id1 = user_id;
     }
+
 
     public String getUser_name() {
         return user_name;
@@ -69,6 +72,7 @@ public class Message implements Serializable {
         this.user_name = user_name;
     }
 
+
     public Long getComment_id() {
         return comment_id;
     }
@@ -76,6 +80,7 @@ public class Message implements Serializable {
     public void setComment_id(Long comment_id) {
         this.comment_id = comment_id;
     }
+
 
     public String getComment() {
         return comment;
@@ -85,13 +90,15 @@ public class Message implements Serializable {
         this.comment = comment;
     }
 
-    public Integer getPost_id() {
+
+    public Long getPost_id() {
         return post_id;
     }
 
-    public void setPost_id(Integer post_id) {
+    public void setPost_id(Long post_id) {
         this.post_id = post_id;
     }
+
 
     public String getPost() {
         return post;
@@ -101,11 +108,12 @@ public class Message implements Serializable {
         this.post = post;
     }
 
-    public Integer getUser_id2() {
+
+    public Long getUser_id2() {
         return user_id2;
     }
 
-    public void setUser_id2(Integer user_id2) {
+    public void setUser_id2(Long user_id2) {
         this.user_id2 = user_id2;
     }
